refactor base class
