+++ pbmohpo/optimizers/eubo.py
@@ -11,11 +11,11 @@ from botorch.models.pairwise_gp import PairwiseGP, PairwiseLaplaceMarginalLogLik
 from botorch.optim import optimize_acqf
 
 from pbmohpo.archive import Archive
-from pbmohpo.optimizers.optimizer import Optimizer
+from pbmohpo.optimizers.optimizer import BayesianOptimization
 from pbmohpo.utils import get_botorch_bounds
 
 
-class EUBO(Optimizer):
+class EUBO(BayesianOptimization):
     def __init__(
         self,
         config_space: CS.ConfigurationSpace,
@@ -29,37 +29,6 @@ class EUBO(Optimizer):
         self.initial_design_size = initial_design_size
         self.new_configs = 0
         super().__init__(config_space)
-
-    def propose_config(self, archive: Archive, n: int = 1) -> List[CS.Configuration]:
-        """
-        Propose a new configuration to evaluate.
-
-        Takes an archive of previous evaluations and duels and proposes n new configurations.
-
-        Parameters
-        ----------
-        archive: Archive
-            Archive containing previous evaluations
-
-        n: int
-            Number of configurations to propose in one batch
-
-        Returns
-        -------
-        CS.Configuration:
-            Proposed Configuration
-
-        """
-        if len(archive.evaluations) == 0:
-            print(f"Running: Intial Design of size {self.initial_design_size}")
-            n = self.initial_design_size
-            configs = self.config_space.sample_configuration(self.initial_design_size)
-        else:
-            configs = self._surrogate_proposal(archive, n=n)
-
-        self.new_configs = len(configs)
-
-        return configs
 
     def propose_duel(self, archive: Archive, n: int = 1) -> List[Tuple[int, int]]:
         """
@@ -123,21 +92,5 @@ class EUBO(Optimizer):
             raw_samples=20,
         )
 
-        configurations = []
-        hp_names = self.config_space.get_hyperparameter_names()
-
-        candidates = [candidates] if n == 1 else candidates.split(1)
-
-        for candidate in candidates:
-            hp_values = candidate[0].tolist()
-
-            config_dict = {}
-
-            # candidate contains only floats, round integer HPs
-            for hp, val in zip(hp_names, hp_values):
-                if not self.config_space.get_hyperparameter(hp).is_legal(val):
-                    val = round(val)
-                config_dict[hp] = val
-            configurations.append(CS.Configuration(self.config_space, config_dict))
-
-        return configurations
+        configs = self._candidates_to_configs(candidates, n)
+        return configs
