+++ pbmohpo/optimizers/optimizer.py
@@ -2,6 +2,7 @@ from abc import ABC, abstractmethod, abstractproperty
 from typing import List, Tuple
 
 import ConfigSpace as CS
+from torch import Tensor
 
 from pbmohpo.archive import Archive
 
@@ -71,3 +72,72 @@ class Optimizer(ABC):
     @abstractproperty
     def dueling(self) -> bool:
         raise NotImplementedError()
+
+
+class BayesianOptimization(Optimizer):
+    def __init__(self, config_space: CS.ConfigurationSpace) -> None:
+        super().__init__(config_space)
+
+    def propose_config(self, archive: Archive, n: int = 1) -> List[CS.Configuration]:
+        """
+        Propose a new configuration to evaluate.
+
+        Takes an archive of previous evaluations and duels and proposes n new configurations.
+
+        Parameters
+        ----------
+        archive: Archive
+            Archive containing previous evaluations
+
+        n: int
+            Number of configurations to propose in one batch
+
+        Returns
+        -------
+        CS.Configuration:
+            Proposed Configuration
+
+        """
+        if len(archive.evaluations) == 0:
+            print(f"Running: Intial Design of size {self.initial_design_size}")
+            n = self.initial_design_size
+            configs = self.config_space.sample_configuration(self.initial_design_size)
+        else:
+            try:
+                configs = self._surrogate_proposal(archive, n=n)
+            except:
+                print("WARNING: Surrogate proposal failed, return random config")
+                configs = self.config_space.sample_configuration(n)
+
+        self.new_configs = len(configs)
+
+        return configs
+
+    def _candidates_to_configs(
+        self, candidates: Tensor, n: int
+    ) -> List[CS.Configuration]:
+        configurations = []
+        hp_names = self.config_space.get_hyperparameter_names()
+
+        candidates = [candidates] if n == 1 else candidates.split(1)
+
+        for candidate in candidates:
+            hp_values = candidate[0].tolist()
+
+            config_dict = {}
+
+            # candidate contains only floats, round integer HPs
+            for hp, val in zip(hp_names, hp_values):
+                if not self.config_space.get_hyperparameter(hp).is_legal(val):
+                    val = round(val)
+                config_dict[hp] = val
+            configurations.append(CS.Configuration(self.config_space, config_dict))
+
+        return configurations
+
+    @abstractmethod
+    def _surrogate_proposal(self, archive: Archive, n: int) -> CS.Configuration:
+        """
+        Propose a configuration based on a surrogate model.
+        """
+        raise (NotImplementedError)
